Add LGTM config Since LGTM.com runs on an old Ubuntu version, I need to build GCC 10 and libstdc++ from source. I need to build it with Clang, since for some arcane reason the GCC devs decided to pass -Wformat-diag as a compiler option, to the building compiler, supported just by GCC >=10, and if you pass an unknown flag to GCC it throws an error. In short, you can only build GCC 10 with GCC 10. Or with Clang, that is smart enough to not implode when you pass it an unkown compiler argument.
But then it came build2, that refused to build itself because
GCC had "CC=clang" in its configuration options,
and it though that I was trying to trick it into compiling itself with Clang
even if I said it that I was using g++.
After 24 hours of madness I decided to
report the bug upstream, build2/build2#136, and it got fixed.

After that Meson started having problems with finding
libodb's pkg-config files, so I needed to modify PKG_CONFIG_PATH.

And then it came simdjson. Ubuntu has an ancient version of the library
that does not have pkg-config nor Findsimdjson.cmake files.
And so I needed to add simdjson as a subproject,
but I kept having problems because I had to explitely say to simdjson
that I wanted a static library, otherwise it would export compiler flags
like -Werror to my own Meson project, causing build failures.

I don't know if this is the definitive lgtm.yml,
because after 3 or 4 days of trial and error I'm getting sick of this,
but it should probably work.

Oh, and I almost forgot to mention that every rebuild took
TWO F*CKING HOURS,
so every time I made a small change to the config I had to
stare into the void untill my cat would ask me to open the door.